fix: rework gemini using asyncio
As gemini 2 only allows one sample at a time (see #4490) I refactored it to use async calls instead. We could also use the batch API, but then it could take 24 hours for a batch (probably wont).

The new script got 0.47 for
'AILAStatutes', with the previous script obtaining  0.47.

I also refactored the different google models into three seperate scripts. No code was changed in the refactor.

Fixes #4490
