+++ mteb/models/model_implementations/google_embeddinggemma.py
@@ -1,0 +1,66 @@
+from __future__ import annotations
+
+import logging
+
+from mteb.models import SentenceTransformerEncoderWrapper
+from mteb.models.model_meta import ModelMeta
+
+logger = logging.getLogger(__name__)
+
+
+MULTILINGUAL_EVALUATED_LANGUAGES = [
+    "arb-Arab",
+    "ben-Beng",
+    "eng-Latn",
+    "spa-Latn",
+    "deu-Latn",
+    "pes-Arab",
+    "fin-Latn",
+    "fra-Latn",
+    "hin-Deva",
+    "ind-Latn",
+    "jpn-Jpan",
+    "kor-Hang",
+    "rus-Cyrl",
+    "swh-Latn",
+    "tel-Telu",
+    "tha-Thai",
+    "yor-Latn",
+    "zho-Hant",
+    "zho-Hans",
+]
+
+
+embedding_gemma_300m = ModelMeta(
+    loader=SentenceTransformerEncoderWrapper,  # type: ignore[call-arg]
+    name="google/embeddinggemma-300m",
+    model_type=["dense"],
+    languages=MULTILINGUAL_EVALUATED_LANGUAGES,
+    open_weights=True,
+    revision="64614b0b8b64f0c6c1e52b07e4e9a4e8fe4d2da2",
+    release_date="2025-09-04",
+    n_parameters=307_581_696,
+    n_embedding_parameters=201_326_592,
+    embed_dim=768,
+    max_tokens=2048,
+    license="gemma",
+    reference="https://ai.google.dev/gemma/docs/embeddinggemma/model_card",
+    framework=["Sentence Transformers", "PyTorch", "safetensors"],
+    use_instructions=True,
+    public_training_code=None,
+    public_training_data=None,
+    training_datasets=GECKO_TRAINING_DATA,
+    similarity_fn_name="cosine",
+    memory_usage_mb=1155,
+    citation="""
+@misc{vera2025embeddinggemmapowerfullightweighttext,
+      title={EmbeddingGemma: Powerful and Lightweight Text Representations},
+      author={Henrique Schechter Vera and Sahil Dua and Biao Zhang and Daniel Salz and Ryan Mullins and Sindhu Raghuram Panyam and Sara Smoot and Iftekhar Naim and Joe Zou and Feiyang Chen and Daniel Cer and Alice Lisak and Min Choi and Lucas Gonzalez and Omar Sanseviero and Glenn Cameron and Ian Ballantyne and Kat Black and Kaifeng Chen and Weiyi Wang and Zhe Li and Gus Martins and Jinhyuk Lee and Mark Sherwood and Juyeong Ji and Renjie Wu and Jingxiao Zheng and Jyotinder Singh and Abheesht Sharma and Divyashree Sreepathihalli and Aashi Jain and Adham Elarabawy and AJ Co and Andreas Doumanoglou and Babak Samari and Ben Hora and Brian Potetz and Dahun Kim and Enrique Alfonseca and Fedor Moiseev and Feng Han and Frank Palma Gomez and Gustavo Hernández Ábrego and Hesen Zhang and Hui Hui and Jay Han and Karan Gill and Ke Chen and Koert Chen and Madhuri Shanbhogue and Michael Boratko and Paul Suganthan and Sai Meher Karthik Duddu and Sandeep Mariserla and Setareh Ariafar and Shanfeng Zhang and Shijie Zhang and Simon Baumgartner and Sonam Goenka and Steve Qiu and Tanmaya Dabral and Trevor Walker and Vikram Rao and Waleed Khawaja and Wenlei Zhou and Xiaoqi Ren and Ye Xia and Yichang Chen and Yi-Ting Chen and Zhe Dong and Zhongli Ding and Francesco Visin and Gaël Liu and Jiageng Zhang and Kathleen Kenealy and Michelle Casbon and Ravin Kumar and Thomas Mesnard and Zach Gleicher and Cormac Brick and Olivier Lacombe and Adam Roberts and Qin Yin and Yunhsuan Sung and Raphael Hoffmann and Tris Warkentin and Armand Joulin and Tom Duerig and Mojtaba Seyedhosseini},
+      year={2025},
+      eprint={2509.20354},
+      archivePrefix={arXiv},
+      primaryClass={cs.CL},
+      url={https://arxiv.org/abs/2509.20354},
+}""",
+    extra_requirements_groups=["embeddinggemma"],
+)
